continue work on restructuring blanim project, progress in blanim.py, there will be multiple types of chain/dag for each crypto/consensus, in some examples you may want to demonstrate invalid/illogical examples.
