+++ blanim/core/__init__.py
@@ -1,0 +1,17 @@
+"""Core base classes for blanim blockchain animations."""
+
+__all__ = [
+    "BaseVisualBlock",
+    "ParentLine",
+    "HUD2DScene",
+    "UniversalNarrationManager",
+    "Frame2DWrapper",
+    "Frame2DAnimateWrapper",
+    "TranscriptManager",
+    # Add other core classes as you implement them
+]
+
+from .base_visual_block import *
+from .parent_line import *
+from .hud_2d_scene import *
+from .dag_structures import *
