+++ blanim/utils/__init__.py
@@ -1,0 +1,20 @@
+# blanim/blanim/utils/__init__.py
+
+"""Optional shared utilities for blanim.
+
+This module contains utilities that may be shared across multiple blockchain
+implementations. Most blockchain-specific utilities should be defined in their
+respective utils/ subdirectories (e.g., blanim/blockDAGs/bitcoin/utils/).
+
+Currently empty - utilities will be added as needed.
+"""
+
+from __future__ import annotations
+
+__all__ = []
+
+# Uncomment when colors.py has content:
+# from .colors import *
+
+# Uncomment when layouts.py has content:
+# from .layouts import *
