+++ examples/__init__.py
@@ -1,0 +1,9 @@
+# blanim/examples/__init__.py
+
+"""Example scenes demonstrating blanim functionality.
+
+These scenes are meant to be run directly via the manim CLI:
+    manim -pql examples/common_examples.py SceneName
+
+They are not intended to be imported as library modules.
+"""
