+++ examples/bitcoin_examples.py
@@ -1,0 +1,1 @@
+# blanim/examples/bitcoin_examples.py
