+++ examples/kaspa_examples.py
@@ -1,0 +1,1 @@
+# blanim/examples/kaspa_examples.py
